// src/modules/ace/interfaces/ace.ts
export interface Ace {
    codigo: number;
    nro_ace: number;
    fecha_registro: string;
    hora_registro: string;
    estado_tabla: string;
    id_ace: number;
    flujo_control: string;
    estado_detalle: string;
    funcionario_asignado: string;
    unidad_seleccion: string;
    numero_unidad_seleccion: string;
    momento_comunicacion: string;
    fecha_hora_comunicacion: string;
    tipo_lugar_inspeccion: string;
    ruc_inspeccion: string;
    razon_social_inspeccion: string;
    direccion_inspeccion: string;
}